+++ test/reflection_factory_test.dart
@@ -487,6 +487,7 @@ void main() {
             TestUserWithReflection,
             TestAddressWithReflection,
             TestCompanyWithReflection,
+            TestFranchiseWithReflection,
             TestDataWithReflection,
             TestDomainWithReflection,
             TestOpWithReflection,